Rename test helper to avoid collection warning
