+++ fsm/tests/test_handlers.py
@@ -1,37 +1,27 @@
-import pytest
-
-from src.core.fsm.handlers import ConcreteStateHandler, ConditionalTransitionHandler
+from src.core.fsm.handlers import ConditionalTransitionHandler
 from src.core.fsm.models import StateStatus
 
 
-def test_allowed_transition():
+def test_allowed_transition(state_handler):
     context = {"current_state": "start", "value": 10, "actions": []}
-    state_handler = ConcreteStateHandler(
-        action=lambda ctx: ctx["actions"].append("processed"),
-        check=lambda ctx: ctx["value"] >= 0,
-    )
     result = state_handler.execute(context)
     assert result.status == StateStatus.COMPLETED
-    assert state_handler.can_transition(context) is True
+    assert state_handler.can_transition(context)
     assert context["actions"] == ["processed"]
 
     transition = ConditionalTransitionHandler(lambda ctx: ctx["value"] > 5, "end")
     outcome = transition.execute(context)
-    assert outcome["allowed"] is True
+    assert outcome["allowed"]
     assert context["current_state"] == "end"
 
 
-def test_disallowed_transition():
+def test_disallowed_transition(state_handler):
     context = {"current_state": "start", "value": 1, "actions": []}
-    state_handler = ConcreteStateHandler(
-        action=lambda ctx: ctx["actions"].append("processed"),
-        check=lambda ctx: ctx["value"] >= 0,
-    )
     result = state_handler.execute(context)
     assert result.status == StateStatus.COMPLETED
-    assert state_handler.can_transition(context) is True
+    assert state_handler.can_transition(context)
 
     transition = ConditionalTransitionHandler(lambda ctx: ctx["value"] > 5, "end")
     outcome = transition.execute(context)
-    assert outcome["allowed"] is False
+    assert not outcome["allowed"]
     assert context["current_state"] == "start"
