+++ pytest.ini
@@ -1,3 +1,5 @@
 [pytest]
-testpaths = tests
+testpaths =
+    tests
+    fsm/tests
 python_files = test_*.py
